OtherUserProfile: refactor ListsContainer component
